sched/ghost: fix never-exiting agents during destroy
If you had an agent task attached to an enclave's cpu and you destroyed
the enclave (e.g. via ghostfs), the agent task would never exit, and the
cpu would be unusable in other enclaves.

The root of the issue is that we would "unpublish" the enclave, meaning
pcpu->enclave = NULL, but the cpu would still be "owned" by the enclave
until the agent exited.  I did this because it was a pain to
synchronize_rcu() in release_from_ghost(), so it was simpler to clear
pcpu->enclave and synchronize_rcu() in ghost_destroy_enclave(), which we
have to do anyways, and then let the agent clean itself up later.

The problem is that the ABI selection uses pcpu->enclave to know which
functions to run.  In this case, we'd set pcpu->enclave = NULL, but the
agent hadn't run yet.  It'd get SIGKILLed, but since enclave == NULL,
PNT wouldn't know *which ABI's* PNT to run, and so we wouldn't run PNT
at all.  Thus the agent would never wake up -> _task_dead_ghost ->
release_from_ghost -> return the cpu.

The fix is to get rid of the split between "publishing" a cpu (mostly
setting pcpu->enclave) and "claiming/returning" a cpu (the cpu_owner
business).  The reason for the split was because synchronize_rcu was a
pain from release_task_ghost, but we can piggyback on call_rcu() from
_task_dead_ghost.

The main concern is that we do three things:
1) pcpu->enclave = NULL
2) sync rcu
3) decref the enclave

call_rcu handles 2 and 3.  I piggybacked on
ghost_delayed_put_task_struct() since I stashed the enclave pointer in
p->ghost.

Note that __enclave_unpublish_cpu() is always immediately followed by
__enclave_return_cpu().  I'll deal with that in a later commit.

Tested: agent_exp, echo destroy > enclave_1/ctl

Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I35c7814eb8bc6b7459897c6e7370170aacadcaf5