sched/ghost: check txn cpu is in an enclave
Since we just checked for an agent task, we should have an enclave.
Eventually, we might drop the need for an enclave to have an agent task
on all cpus.  This will cover us if we ever remove that agent task
check.

Tested: edf_test, enclave_test, api_test
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I3eee6303d126404e25fa079be1cc28a4cd12affb