sched/ghost: add warnings to catch kernfs_mutex deadlocks
There's no way to prevent ghost client tasks from grabbing the
kernfs_mutex.  It's used in sysfs and other places.

If an agent task, responsible for scheduling tasks, ever attempts to
grab the kernfs_mutex, there's a chance we'll deadlock.  Specifically,
the client can hold the mutex and be descheduled.  If the agent blocks
on acquiring the mutex, it will never schedule the client.

In general, ghostfs operations on open FDs don't touch the kernfs_mutex
- there might be corner cases I didn't see.  But we definitely grab it
when opening paths (walking or lookups).  The agent doesn't currently
open any files from ghostfs from agent tasks, so this warning hasn't
fired yet.  But at least we'll catch it if it does and avoid a potential
source of deadlock.

Tested: edf_test, enclave_test, api_test
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I38422580ccc6b098fd2ee556d83e09f919a467cb