sched/ghost: allow lockless requests to resched prev
Currently, whether or not prev is allowed to be selected by PNT is
controlled by rq->ghost.must_resched, which is protected by the RQ lock.

Setting must_resched means that the next time PNT runs, prev will be
preempted by something - even if it is the idle task.  (When prev is not
an agent).

An agent can request that prev be unscheduled by issuing a NULL
transaction.  This grabs the RQ lock.  We will need to preempt tasks
from BPF-MSG, which already holds an RQ lock, so we need to set
must_resched, or something similar, locklessly.

The lockless nature of this is tricky - if we're in PNT, already ran
BPF-PNT, and then the agent tries to resched the cpu, do we want the
resched to take effect or not?  Was the agent referring to prev (who is
different than the latched task)?  Or referring to the task they just
latched?

This sounds a lot like why we have barriers.  The agent wants to change
state, based on its understanding of the system.  Only apply that change
if the agent's understanding matches reality.

We already have the cpu_seqnum: the history/state of a given cpu,
exposed in all relevant messages.  Even better, unlike the barriers,
userspace has no assumptions about being told about every increment to
cpu_seqnum.  We can increment it whenever any cpu_state changes and we
don't have to send a message.

So when the agent (via bpf, in an upcoming commit), asks us to resched,
they can tell us the cpu_seqnum they got from their last TASK_LATCHED.
As a side note, this means that they should *not* use the return value
of bpf_ghost_run_gtid() to mean "it's ok to resched right now!", since
they don't know the cpu_seqnum.  Even if we returned that value, as soon
as we send TASK_LATCHED, we increment again (it's another change in cpu
state, perhaps better renamed as TASK_ON_CPU, but that's a separate
issue).  The agent should only resched based on the latched->cpu_seqnum
from when prev got on cpu.

Tested: edf_test

Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ie9edaab22ae2f77078ca2b887fe72d4d528b2e80