sched/ghost: add bpf_ghost_resched_cpu
This is a bpf helper, callable from any enclave BPF program (i.e.
BPF-PNT, BPF-MSG, or skip-tick).  It will force the cpu to reschedule,
such that "prev" will not be picked again.

Agents can use this to preempt a task.  When the cpu reschedules,
assuming there is no latched task, BPF-PNT will run so that the agent
can pick another task to latch.

A few notes:
- You can call this from within BPF-PNT.  If you call it on your own
cpu, it will have no effect, mostly, similar to setting need_resched
during pick_next_task().  The technical reason it won't work is that the
kernel calls clear_preempt_need_resched() right after pick_next_task(),
so any "need_resched" set during PNT will be a noop.  "Mostly", since
you could write "cpu_seqnum + N" and have must_resched set at some
arbitrary point in the future.  (The kernel cleans this up during
cpu/agent teardown).

- check_same_enclave() works so long as you are in an rcu read critical
section.  We're a bpf helper, and our bpf programs are run under an
rcu_read_lock().  Someone could remove the cpu from the enclave, but
that cpu cannot be handed out to another enclave until after a grace
period.  This is the case for all of the ghost bpf helpers.

- I think you don't need an smp_wmb() between the WRITE_ONCE and the
resched_cpu_unlocked() (which is a prodkernel function).  That has a
CAS in it, and in general, I'd expect sane behavior between the
ordering of writes and IPIs in Linux.  Though that might not be sane.

To be submitted with cl/422883910.

Tested: edf_test, using the helper in bpf-msg.
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I2746e6c789010c3a1b0014483d3d2b788723149f