sched/ghost: fix bpf_helper enclave discovery
Certain operations, particularly ghost bpf helpers, are operating on
some enclave, but it might not be the enclave that owns the cpu.

Consider an enclave on a set of cpus, and a task on another cpu enters
the enclave.  That generates a TASK_NEW, which triggers BPF-MSG.  That
all runs on a cpu that is not in the enclave.

Add per-task *__target_enclave to track which enclave we are targeting
with our functions.

Users of the target_enclave can nest, such that if we are in the middle
of some operation and get interrupted, the IRQ handler can set its own
target_enclave so long as it restores it.

Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I1cf7bc70bf2bda6d133155b9a8877ca531074587