sched/ghost: use ghostfs to move tasks to an enclave
Instead of using syscalls, use ghostfs to move tasks to an enclave.
This cleans up core.c's setsched code.  We had a weird dance of passing
along a ctl_fd, then converting that to an enclave, but we had to put
the enclave after unlocking the RQ.  It wasn't pretty.

Instead, if we enter the kernel through ghostfs, the enclave is already
known, and its refcounts are managed for us.  This is analogous to using
ghostfs ioctls instead of raw syscalls: we do not need any magic to
figure out which enclave we're using.

To that end, there are two mechanisms to enter ghost, one for regular
tasks and another for agents.

To add a regular task to an enclave, write its pid into enclave_x/tasks.
Write 0 for 'current'.  The tasks file is world writable, so anyone can
join an enclave, if they can find it.  However, to move another task
(pid != 0), you need CAP_SYS_NICE, just like with
sys_sched_setscheduler().

You can read the tasks file to get a list of tasks in the enclave,
though beware that if you have too many tasks, the file might be
truncated (seq file overflow).

To add an agent to an enclave, write "become agent <QFD>" into ctl.  You
need write access to ctl, which an agent will have.  The QFD is the FD
of the queue you want for this agent, -1 for "default".  qfd is the
same parameter we used to pass in the sched_attr.

To be submitted with cl/424409182.

Tested: edf_test, enclave_test, api_test, transaction_test,
manually moving tasks into an enclave from the shell.

Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ib4ead6536354259ca1a464e62d72220c0465eef4