sched/ghost: split out bpf helper ID definitions
enum bpf_func_id is created by a macro that is used for an array lookup
from helper ID to string (func_id_str in disasm.c).

We don't want that array to be too big.  However, we also don't want our
helper ID numbers to clash with upstream's function IDs.  That will
happen eventually.

Move the ghost bpf helper IDs out of bpf_func_id so that we can use a
larger number without affecting the array.

The array mapping is optional - it's used in func_id-to-name lookups,
and those lookups are protected in case the func_id > __BPF_FUNC_MAX_ID.

Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I5b221d1834f1afe11aed55f73216b9e3d8915f46