sched/ghost: increase bpf magic numbers
With the growing number of BPF helpers, we'll soon run out of space in
the bpf_func_id space.

Increase the base numbers for PROG_TYPEs, the ATTACH_TYPEs, and helpers:
- progs:   1000
- attach:  2000
- helpers: 3000

There is enough room for growth in upstream's numbers that we'll either
be merged upstream or dead before we run into conflicts.  We're still
well under the 0xffff requirement for attach types (since the upper bits
are the enclave ABI).

To be submitted with cl/424689285.

Tested: edf_test
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ie0600e0b4478d3289377e4a4d58c9e13b4a491aa