sched: poisoned rendezvous should not produce TASK_PREEMPT.
Prior to this change a task with a poisoned rendezvous could
produce a MSG_TASK_PREEMPT (e.g. if the next task picked to
run on the cpu belongs to CFS). This is not intuitive since
the overall sync_group has failed. It is also easy to miss
since producing MSG_TASK_PREEMPT is the exception. For e.g.
see cl/427576696 that fixes flaky ApiTest.LocalAgentWakeup

Effort: sched/ghost
Google-Bug-Id: 214648944
Change-Id: Ib13c38f60b8d1acbd292d4de7b8d7ff528f44c4f