sched/ghost: always use select_task_rq_ghost()
select_task_rq() did not always call select_task_rq_ghost(), and even if
it did, it could ignore the answer - particularly if the cpu was not in
the task's allowed_cpus.

The RQ doesn't matter much for ghost.  The "real" RQ is in the agent,
but the kernel still uses the (per-cpu) RQ for its own bookkeeping,
including the steps needed to wake a task.  For an in-kernel scheduler
like CFS, the RQ matters: it's where the task will compete for cpu time
and may ultimately run.  For ghost, it's a temporary staging ground
until a task gets latched (at which point it will change RQs).

Of note is the TTWU_QUEUE feature.  Normally, when we do a 'remote'
wakeup, meaning select_task_rq() picks a cpu other than the caller's,
we'd grab their rq lock, do the wakeup, and maybe IPI if we need to
preempt.  With TTWU_QUEUE, we can put the task on the RQ's wake_list,
then send an IPI (amortized).  This trades an IPI for lock bouncing -
presumably if the task will actually run on its old cpu, then this will
help.

The problem arises when we pick a rq that is neither the task_cpu (i.e.
previous cpu it ran on) nor the current cpu.  This third-party cpu can
get overloaded with resched IPIs from TTWU_QUEUE remote wakeups.

CFS tries to find an idle RQ, and doesn't aggressively load balance or
otherwise migrate tasks from a cpu.  Ghost does neither.  We offer a
choice between task_cpu and the current cpu (waker's cpu).  And we can
yank tasks off remote RQs without hesitation: this is what happens
during a latch.

select_task_rq_ghost() would normally be OK, but prior to this commit,
select_task_rq() would ignore the result under certain circumstances,
specifically if the cpu was not in p->allowed_cpus.  The ghost kernel
ignores affinity: that's up to the agent to enforce, yet
select_task_rq() was enforcing affinity, but only for the 'temporary'
rq - not the actual execution of a task.

Consider a task that was affined to numa node 1, yet the agent ran it on
node 0.  It wakes on node 0, and select_task_rq_ghost() returns a cpu
from node 0.  select_fallback_rq() will pick the first cpu on node 1 to
do the wakeup, and using TTWU_QUEUE, it will send an IPI and that cpu
will have to do all the wakeup bits.  For ghost, that includes sending
the wakeup message, which can run a bpf-msg program.

Once the task is woken up, another cpu (perhaps from node 0) can latch
it - essentially pulling it away from the victim cpu.  If that task
blocks and wakes quickly, and it was running on node 0, the cycle will
continue.

If you have enough cpus running/blocking/waking tasks, and shipping
their wakeup work to a victim cpu, then that cpu will get stuck in IRQ
context, constantly handling new resched IPIs.  You'd need enough cpus
to keep the victim busy.

The aggrivating factor to this is bpf-msg.  bpf programs will finish in
a bounded amount of time, but there's nothing that stops them from being
inefficient.  The world's simplest scheduler (my Biff scheduler in a
forthcoming CL) is a global queue, implemented with a
BPF_MAP_TYPE_QUEUE, which is protected by a kernel spinlock, which is an
MCS queue lock.  If you have all the cpus hammering that lock too, it
will make bpf-msg less efficient, to the tune of 40-50us per task
ttwu().

Even without bpf-msg, you'd need only enough cpus to keep the victim
busy.  If the wakeup cycles is W_c, and the time to block, wake, run,
and reblock is T_c (task), you'd need T_c / W_c cpus to keep the victim
stuck in its IPI handlers.  But a combination of bpf-msg and a really
dumb scheduler increased W_c.

The trick is to break the cycle: by requiring that select_task_rq()
returns either task_cpu() or the current cpu, we avoid this "stuck in
resched IPI" scenario:
- if it's the current cpu, we don't need an IPI.
- if it's task_cpu, then when some other cpu pulls it from the victim's
  RQ, that cpu becomes the new task_cpu, breaking the cycle.

Tested: SLL
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ib65a1672420c72ff9d3bf71c0d50eba245a1019d
Origin-9xx-SHA1: 37588684a0ce59fb763a4deb2fea31078e47984c
Rebase-Tested-11xx: compiled