sched/ghost: always force_offcpu() in _set_curr_task_ghost()
Whenever we call set_curr_task(), we need to kick the task off cpu.
Waking the agent on that cpu usually does that - assuming the agent is
on_rq and can be woken.  That is currently the case, since PNT-G aborts
of the agent isn't on_rq, but that shouldn't be necessary.

Furthermore, waking the agent is overkill.  The agent task could be a
satellite agent, even though we want the global agent to respond to the
change (via some MSG).  The agent task could also be from a different
enclave.  Consider the case of a CFS task joining ghost: it can do so
from any cpu in the system.

The simplest is to just force_offcpu(), which is just setting
must_resched to avoid this task from being picked again in PNT-G.  We'll
get a preemption message (sent to the correct agent!) due to
produce_prev_msgs().

Tested: api_test, transaction_test, edf_test
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ide312b54905bf7fcf3d3db881eebbbb92ba02ded
Origin-9xx-SHA1: f894a3b1a7f4ce8728b6108407aab95451189710
Rebase-Tested-11xx: compiled