sched/ghost: skip the CFS PNT optimization if ghost is active
Essentially, we're trying to determine if we should run PNT-A and
PNT-G without actually running them in a low-overhead manner:
- The agent will want to run if it's on_rq and not blocked_in_run.
- The agent may want to run a task if there are any tasks in the
  enclave (latched or a pending txn)

Added wrinkles:
- We might be okay with the optimization calling PNT-F immediately
  (say the agent is blocked_in_run), but we won't want to call PNT-Idle
  (say we have a task in the enclave), which is the second half of the
  optimization below.
- This rq->nr_running doesn't know about tasks in the enclave on
  other rqs.  In lieu of checking the enclave's run list (ABI
  dependent), we can check for the existence of an agent to imply the
  potential for some task in the enclave.  (Yes, that's somewhat ABI
  dependent, but we already have references to the agent elsewhere in
  core.c).
- If the agent is on_rq, it counts towards rq->nr_running.

We could allow the optimization to call PNT-F if we detect the agent is
blocked_in_run, (adjust nr_running, etc) but then we'd want to skip the
PNT-I call if there is even a chance of there being a task in the
enclave.

It's also not clear that we even want the PNT-F optimization if there is
an agent.  Consider the case where this rq has no ghost tasks
(non-agent).  We'd run PNT-F, fail, then run the entire PNT loop from
the top, including PNT-F again.

This all isn't an issue yet, but once we allow ghost tasks to run when
the agent is not on_rq, we'd start hitting this optimization, and we'd
potentially strand latched tasks until the agent is back on_rq.  (Task
is latched, agent is !on_rq so nr_running == cfs.h.nr_running, so we'd
never call PNT-G).

Disabling the optimization only when we have an agent will at least
cover us for having the ghost kernel in the fleet, but with no agent
loaded.  Which was the status quo.  The difference this commit makes is
that when the agent is not on_rq, we will still disable the
optimization.  Previously, we'd only disable the optimization when the
agent was on_rq and contributing to nr_running.

Tested: api_test, transaction_test, edf_test
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ia20d187814c5d4ed54684805d23bf513039921b0
Origin-9xx-SHA1: 82fd823820db8ddb00c29fdc174f57c5068471b9
Rebase-Tested-11xx: compiled