sched/ghost: don't leave ELIDE_PREEMPT set after use
ELIDE_PREEMPT was set ghost_set_pnt_state(), along with latching a new
task: the rq run_flags *correspond* to the latch_task.  It means that
when latching this task, if there was a previous task that gets
preempted, *don't* send the message.

However, once we do this, we shouldn't leave the flag set.  We'll
eventually run the latched task (in PNT-G at "pick_latched:").  We do
*not* want to elide the preemption for that latched task if we see a
stale run_flags.

If a task latches and sets ELIDE_PREEMPT, that should have only applied
to when it got on cpu, and applied to whoever was prev at the time.

Similarly, if the task was latched, but we don't actually pick the task,
e.g. a task departs ghost or we migrate it to another cpu, we don't want
that ELIDE_PREEMPT flag to linger.

Note: it's always correct to send a preempt, i.e. to ignore the
ELIDE_PREEMPT.  CFS can activate a thread at any time, so the agent must
be able to handle a preemption msg, even if it didn't want to get one.

Tested: api_test, edf_test
Effort: sched/ghost
Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I80307231e9cf20687d099307ade130c77df10a46