sched/ghost: remove cpus from enclaves with a two-step method
We use RCU to protect pcpu->enclave.  However, we were "returning" the
cpu to the global system in ghost_remove_cpu(), but had no protection
from reusing the cpu.  A while back we had the two step: unpublish,
sync_rcu, then return the cpu.  Return to that method of removing a cpu
from the enclave.

The issue becomes more apparent when we rename the old "return" cpu to
what it really is: ghost_free_cpu().  This is a classic deallocating an
RCU-protected pointer problem, and you don't free the object until you
sync_rcu.

In particular, the "object reuse" applies to anyone who uses RCU to get
the pcpu->enclave pointer: how did we know the cpu still belonged to the
enclave?  We didn't!  The old code protected the *lifetime of the
enclave struct*, but not the membership of the cpu in the enclave.  And
that cpu could be reused concurrently.

And there's actually a data reuse/clobbering aspect to this too!
Consider pcpu->ghost_txn.  This is an RCU-protected pointer to
&pcpu->enclave->cpu_data[cpu]->txn.  We're just caching the lookup to
avoid chasing a half dozen pointers.  No problem - it's all in the same
enclave.

Now consider rq->ghost.  Many of those fields are *essentially* a
(nonexistent) struct pcpu->enclave->ghost_cpu_blob[cpu].  We use
rq->ghost as a bunch of enclave per-cpu fields.  E.g. must_resched,
ignore_prev_preemption, etc.  That's all ABI-dependent too!  There are
other fields in that struct that are ABI and enclanve-independent, e.g.
the enclave_work stuff.  Or at least that should be enclave independent,
via a stricter API, etc.

Point being: if we think we had an RCU-protected pcpu->enclave pointer,
we thought we were allowed to muck with rq->ghost.  But that data
structure is "freed", and can be reused by another enclave!  Example:
bpf_ghost_resched_cpu().  We checked the enclave, then just started
poking rq->ghost and sending IPIs!  Note that function doesn't
explicitly rcu_read_lock() yet, though BPF runs under and RCU read lock.

The fix is to explicitly free the cpu, after a synchronize_rcu().

No need to incref the enclave when deferring the cpu-free to the agent's
death.  (the old __agent_decref_enclave).  The existence of the agent
keeps the enclave alive, via p->ghost.enclave, which is dropped after we
free the cpu.  And even if we didn't, we don't actually need the enclave
struct to free the cpu.  I'm not 100% sure why we had that incref in the
first place.  I think it was to make sure that the enclave itself wasn't
freed until after a sync_rcu, which is about protecting the enclave
struct itself (which we were doing correctly before.  This commit is
about protecting the enclave-cpu mapping).  But now we don't actually
refer to the enclave in ghost_delayed_put_task_struct().

Change-Id: I84dc8e5735b37345618abce0fb8ddc772b9430cc