sched/ghost: use RCU to protect the cpu allocation in __ghost_run_gtid_on()
Technically, since we were called from bpf, we were already under an
rcu_read_lock().  Make the use of RCU to protect the cpu's allocation to
the enclave explicit.

Additionally, we always check the enclave now.  Currently, we're only
ever called from PNT, and the target cpu is always the current cpu, but
the check is harmless and may catch future bugs.  Particularly, if we
ever get rid of the requirement to have an agent task, we want some way
to notice if a cpu is being yanked out of an enclave.

Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Iee1c0f43e6cae28b44b6fd9c8ef360989b293930