sched/ghost: remove the percpu(enclave) check in _ghost_commit_txn()
The check was racy an unnecessary.  We're already operating under and
RCU lock, and we have the ghost_txn pointer, which has the same
guarantees as per_cpu(enclave, cpu_of(rq)): if we saw it with an
rcu_read_lock(), then the cpu has not been freed / returned to the
system.

It's possible for us to have seen a ghost_txn pointer but not see
pcpu->enclave.  That's fine - that just means we caught it in the middle
of "unpublishing" the cpu mapping.  It's just dumb luck whether we see
it or not, since the unpublish could happen right after we did our
check.  The check is just a false sense of security.

Note that we still have to check for an agent - we don't want to latch a
task and have it stranded or anything.

Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I618d8cf3d5c3739a039ba77d44a89e6cd7c49c06