sched/ghost: add MSG_CPU_AVAILABLE and MSG_CPU_BUSY
Whenever a cpu changes between available and busy, generate a message,
controlled by the enclave tunable deliver_cpu_availability (default
false).

The kernel exports the availability of a cpu via the status word table.
Userspace agents can poll the table, and perhaps they do not need to
be notified of changes in cpu availability.

Arguably, the design of the status_word's CPU_AVAIL bit was meant to cut
down on the number of cpu availability messages; on a CFS-busy system
the agent could get flooded with messages.  But BPF handles messages
quickly and synchronously, so a high message rate doesn't matter as
much.  Though even userspace global agents with a tight loop can use
MSG_CPU_AVAILABLE to keep an up-to-date mapping of their available cpus
instead of scanning the status word table.

Additionally, with BPF, you could implement your own agent-specific
status_word table as a bpf map.  See biff_bpf.h for an example in the
accompanying change.

Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I3acfff6fab1becca76bb92c905a04208feef5fcb