sched: Fix txn state update vs. TASK_PREEMPT race
This KCL fixes an issue where a TASK_PREEMPT message is delivered for a
scheduled task *before* the txn that scheduled the task has its state
updated to GHOST_TXN_COMPLETE. This issue is triggered by running
simple_exp over and over again and having the threads scheduled by
agent_sol. Generally, the TaskDepartedManyRace test will fail after 5-10
minutes of repeated runs.

This issue is triggered by the race below, when a global agent schedules
a remote CPU with an asynchronous txn.

1. The global agent is running on CPU B. The global agent wants to
   schedule CPU A with commit_flags == 0. Thus, when CPU B is finished
   filling in the txn, it sends a resched IPI to CPU A, and it is now
   CPU A's responsibility to commit the txn.

2. CPU A receives the resched IPI and calls: scheduler_ipi() ->
   ghost_commit_greedy_txn() -> _commit_greedy_txn() ->
   _ghost_commit_pending_txn() -> ghost_commit_txn() ->
   _ghost_commit_txn()

3. CPU A acquires the rq lock in _ghost_commit_txn(). After the commit,
   the task is latched on CPU A and CPU A drops the rq lock again, *but
   before writing the txn state*.

4. _ghost_commit_txn() returns back into ghost_commit_txn().
   ghost_commit_txn() writes the txn state.

In between steps (3) and (4), someone on some CPU C (where C != A) moves
the task out of ghOSt. CPU C calls _ghost_setscheduler() which calls
invalidate_cached_tasks(). invalidate_cached_tasks() calls
ghost_latched_task_preempted() which sends a TASK_PREEMPT message to the
global agent and unlatches the task on CPU A.

The global agent sees the TASK_PREEMPT message before step 4 happens
above, resulting in a failed CHECK since the txn is still pending.

This seems like a very difficult race to trigger, and it is for the most
part.  However, one scenario particularly susceptible to this race is
when we call ghost_commit_greedy_txn() from should_resched(), and
should_resched() may be called while IRQs are enabled (such as via
cond_resched()). Thus, once the rq lock is dropped in
_ghost_commit_txn() with rq_unlock_irqrestore(), it is possible that a
pending interrupt is immediately delivered to CPU A. Thus, CPU A spends
time servicing the interrupt before writing the txn state, making this
race easier to trigger. The other cases (e.g., the scheduler_ipi() case)
are less susceptible to this race because they already have IRQs
disabled, so dropping the rq lock with rq_unlock_irqrestore() does not
turn IRQs on.

------

It is also possible to trigger this issue if a global agent schedules a
remote CPU with a *synchronous* txn (i.e., COMMIT_AT_TXN_COMMIT is set).
In that case, it is possible for the global agent CPU to commit the txn
and send the resched IPI, and then start handling an interrupt once the
rq lock is released but before the updating the txn state. Then while
the global agent CPU is handling the interrupt, the task on CPU A is
preempted and so a TASK_PREEMPT message is delivered (or the task could
block, yield, etc., and generate those messages instead).

------

Change-Id: I44bdadc73a6fac25f7029009840d66a48bcb4432