sched/ghost: add MSG_CPU_AGENT_BLOCKED and MSG_CPU_AGENT_WAKEUP
Whenever an agent blocks or wakes up, such as due to a futex or page
fault, generate a message, controlled by the enclave tunable
deliver_agent_runnability (default false).

"Voluntary blocks", a.k.a. "blocked_in_run" do not count as block/wakeup
edges.  Those blocks are when the agent explicitly parks itself in the
kernel, awaiting a ping or letting some other task run.  In those cases,
the agent task is runnable and on_rq, but not selected to run.

Userspace agents can't do too much with this, as-is, depending on which
Channel gets the message.  You'd have to route the message to another
agent, since the given agent is blocked.  Though an agent could find out
it blocked after the fact.

BPF can make use of this info, performing actions when it detects that a
cpu's agent is offline.  e.g. BPF can wake the sibling cpu's agent to
enforce a Core Scheduling policy.

Aigned-off-by: Barret Rhoden <brho@google.com>
Change-Id: Iab34de6e0f4e682d32090abd726eab4e44956a41