sched: update 'txn->state' for sync_group cpus as each txn is committed.
Prior to this change we left 'txn->state' alone in _ghost_commit_txn()
if the transaction was part of a larger sync_group commit. We did this
because the semantics we guaranteed were as follows:
- on a successful sync_group commit all transactions in the group will
  have txn->state = GHOST_TXN_COMPLETE.
- on a failed sync_group_commit one of the transactions in the group
  must have failed with the appropriate txn->state and the rest will
  have txn->state of GHOST_TXN_POISONED. Note that the latter category
  is composed of earlier transactions that succeeded but were later
  revoked (via a poisoned rendezvous) or those that were never even
  attempted.

These semantics are not particularly useful since the return value
of ghost_sync_group() itself indicates whether the overall commit
failed or succeeded.

We now change this such that each individual commit updates its
'txn->state' in _ghost_commit_txn().

Change-Id: I5fedbfc60c1a2c80b43392db783bb30b512d4efb