sched/ghost: return success for idempotent joins to an enclave
When moving a task to enclave, which is via writing to the enclave tasks
file, previously we returned EPERM if the task was already in ghost.
The reason for that is we don't want to handle corner cases of switching
enclaves.  It keeps the state machine simpler.

However, if you're already in the enclave you are trying to join, we can
detect that and return 'success' (the length of the write() is success).
This is similar to setscheduler: if you're already in the scheduler,
that is a success.

For clients trying to join an enclave, it's simpler for them to say "put
me in ghost, and if i was already there, that's fine", instead of having
them handle an error case.  This is especially true if the client
detected the agent changed (i.e. agent_online toggled), but the thread
never left the enclave, which is what happens during an agent live
update.

Change-Id: Ia8e1db756cfb035329d448c3e2b02fb788afc14f