sched/ghost: use a helper to handle cpu availability
Whenever the cpu availability changes, send messages and change the
status word (SW).

handle_cpu_availability() is idempotent: we can call it as many times as
we want, from wherever we want (with rq lock held!).

This is a slight change in functionality: previously, every context
switch, we'd write to the status word, even if the availability had not
changed.  Now, only set/clear the SW flag when we notice a change.

Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ib3f04daeef88443f7583f0f5990a59ec26f06102