sched/ghost: send MSG_CPU_AVAILABLE before BPF-PNT
Consider a BPF agent assigning tasks to available cpus.  One strategy
would be to hook MSG_CPU_AVAILABLE and assign a task, then latch it in
BPF-PNT.  However, if we don't send Availability until
prepare_context_switch, we'd run BPF-PNT before knowing about
availability.  What should a bpf-agent do then?  It'd have to pick some
background work or just allow the cpu to idle, and then wait for
MSG_CPU_AVAILABLE, and then resched itself to pick the task it actually
wanted.

Instead, send AVAILABLE before BPF-PNT in the common case.
handle_cpu_availablility() is idempotent.  We can call it at any time,
and the sooner the better.  If it turns out that CFS activates on the
CPU after we send AVAILABLE, we'll catch it in prepare_context_switch()
and send BUSY.

We still must call handle_cpu_availability() in
prepare_context_switch(), i.e. after PNT makes its choice.  It's
tempting to only check in pnt_prologue, but the availabilty depends on
the result of PNT.

The more obvious case is AVAILABLE->BUSY, where we see the CPU is
available in pnt_prologue(), and after we unlock the RQ in PNT, a CFS
task lands on this cpu and we ended up picking it in PNT.  We'll catch
that after PNT, in prepare_context_switch().

The other case is BUSY->AVAILABLE.  We could see the CPU being busy in
pnt_prologue().  Consider a CFS task is on_rq, so we're busy (regardless
of whether we just became busy at this moment and sent a BUSY message,
or if we were always busy).  BPF-PNT ran, thinking that the cpu was
busy.  During PNT, we unlock the RQ and the CFS task migrates away.
We'll likely pick the idle task in PNT.  prepare_context_switch(), when
we switch to idle, will notice that the CPU was busy but is now
available, and will send MSG_CPU_AVAILABLE.  (at which point, bpf can
set need_resched()).

The latter case is essentially what we do prior to this commit: have an
extra context switch each time we go from busy to available.  By sending
AVAILABLE as early as we can, we make this context-switch case
infrequent.

Side notes:
- For blocking tasks, e.g. the last CFS task on a cpu blocks, rq
  nr_running is decremented before pick_next_task() in __schedule() via
  deactivate_task(), so the check in cpu_is_available() works in
  pnt_prologue().
- We probably don't need ctx->might_yield anymore in bpf-pnt.  BPF
  schedulers can use MSG_CPU_AVAILABLE and BUSY.  Though might_yield is
  a convenience for the short term, and maybe for dumber schedulers.
  Once we fix how idling works, maybe might_yield and dont_idle can both
  go away.

Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: Ifeb8393d52aa7bc160663259ae0cec862b6c11d2