sched/ghost: check sched_class during bpf_ghost_resched_cpu()
It was possible for a BPF program to spam IPIs to cpus running CFS or
higher classes.

The agent can perpetually ping a cpu too, but that is slightly
different: that wakes the agent (which BPF does with
bpf_ghost_wake_agent()).  We can add safety checks to avoid agent cpu
monopolization, but it's not clear that that would cover the case of a
cpu constantly being IPIed, where the agent doesn't have to run.

Signed-off-by: Barret Rhoden <brho@google.com>
Change-Id: I8b70e0c919bee0e530119fe905e8ffb67c13a8f5