sched: resync agent barrier when agent preempts ghost task
The path for an agent preempting a ghost task, from e.g. task tick,
looks like:

- task_tick -> bump the agent barrier and resync it
- agent is picked -> bump the agent barrier due to preempting a
  ghost task

The issue though is that after the second agent barrier bump, we don't
resync it with the status word. As a result, the next local transaction
committed by the agent will fail with AGENT_STALE, forcing a retry. We
can avoid this spurious commit failure by resyncing with the status word
barrier after bumping the barrier for the ghost task preemption.

Change-Id: Ia5d0e76af1bdc59564e1f29654cae54136816028