sched: add static_key to check whether ghost is active
Adds a static key to determine whether ghost is in use at all. This
allows us to shortcut paths that would normally be executed when ghost
is compiled into the kernel but inactive.

Two particular areas we can shortcut are ghost_commit_greedy_txn and
ghost_cfs_added_load, which each consume a meaningful amount of cpu even
when ghost is not active.

Change-Id: I8faae79d3d1e7e77c31996909b8fe171eae042d9