sched/ghost: remove WARN_ONs that are not kernel bugs
These warnings were triggerable by userspace and were often used for
error cases.  WARN should be used only for kernel bugs.

The most popular of these was overflowing a ghost_queue, and without a
print, we wouldn't know we dropped a message.  In other cases, such as
ghost_cpu_data_mmap(), we were propagating an error message to userspace
already, so there is no need to print to the kernel log.

Change-Id: I552fd3e35a6704512981c44faf89cc1de2b128a0