sched/ghost: track the caller's state in BPF programs
Previously we were using rq->ghost.in_bpf_pnt to track whether we were
in BPF-PNT or not.  More broadly, we want to know which program type we
are in, and also pass arguments for that type, which we'll use in
upcoming patches.

Note that BPF programs are non-preemptible, so we can use this_cpu.
That's a little cleaner than cluttering rq->ghost, which we were
somewhat using as per-cpu storage.

Also note that we're currently using rq->ghost as a communal per-cpu
dumping ground for all of the ABIs.  If the cpu is in your enclave, then
your ABI can do what it wants with thos variables.

Change-Id: I544be89897c1117ab57b9ef0e7410c3ecadf3e16