sched/ghost: hold the RQ lock during BPF-PNT
Previously we'd drop the lock so that we can grab any RQ lock during
ghost_run_gtid.  However, it may be the case that we do not need to drop
the lock at all, particularly if the task is already on this cpu.

Using the ghost_bpf_prog_state, we can communicate certain facts to our
BPF helper, such as the RQ's 'rf' (lock "pinning" state, largely used
for debugging, as well as convenient storage for irq_flags for
task_rq_lock()).

Additionally, since we return to BPF with the RQ lock held, we can now
guarantee to the BPF agent that we continue to hold the RQ lock from the
moment the task was latched all the way back into BPF.  This is helpful
for cases where we had to drop the RQ lock to migrate the task to this
cpu.  The result of a successful latch is synchronized w.r.t.  all other
messages about the task by the task rq lock.

Note that BPF was never allowed to ask for a remote latch, but the
internal interfaces allowed it.  Remove those - if we need a remote
latch from BPF, it won't have the same guarantee.  We hold *this* cpu's
RQ lock when we return to BPF, but if we latched the task on another
cpu, we won't hold the task's RQ lock when we return.

Change-Id: If1b9c06c7bd947d78da9a41cd618c0d6fb04c456